+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.theme.service.dto.AvailableTimesResult;
 import roomescape.time.domain.ReservationTime;
+import roomescape.time.exception.DuplicateTimeException;
+import roomescape.time.exception.TimeNotFoundException;
 import roomescape.time.repository.ReservationTimeRepository;
 
 @Service
@@ -19,12 +21,20 @@ public class ReservationTimeService {
 
     @Transactional
     public ReservationTime registerReservationTime(ReservationTimeCommand command) {
+        if (reservationTimeRepository.existByStartAt(command.startAt())) {
+            throw new DuplicateTimeException();
+        }
+
         ReservationTime reservationTime = ReservationTime.of(command.startAt());
         return reservationTimeRepository.save(reservationTime);
     }
 
     @Transactional
     public void removeReservationTimeById(Long id) {
+        if (reservationTimeRepository.findById(id).isEmpty()) {
+            throw new TimeNotFoundException();
+        }
+
         reservationTimeRepository.deleteById(id);
     }
